Move test utils for graphql common plugin to separate folder
